refactor: move shared sub command under pkg, don't include the connector service in the kas-fleet-manager binary
we also remove it's cli flags from the service template.
