+++ cmd/kas-fleet-manager/main.go
@@ -4,7 +4,6 @@ import (
 	"flag"
 	"github.com/bf2fc6cc711aee1a0c2a/kas-fleet-manager/internal/kafka"
 	"github.com/bf2fc6cc711aee1a0c2a/kas-fleet-manager/pkg/environments"
-	"github.com/bf2fc6cc711aee1a0c2a/kas-fleet-manager/pkg/providers/connector"
 	"github.com/golang/glog"
 	"github.com/spf13/cobra"
 )
@@ -24,7 +23,6 @@ func main() {
 
 	env, err := environments.NewEnv(environments.GetEnvironmentStrFromEnv(),
 		kafka.ConfigProviders(),
-		connector.ConfigProviders(true),
 	)
 	if err != nil {
 		glog.Fatalf("error initializing: %v", err)
