--- templates/connector-catalog-configmap.yml
@@ -1,35 +1,0 @@
----
-kind: ConfigMap
-apiVersion: v1
-metadata:
-  name: connector-catalog-camel-aws
-  annotations:
-    qontract.recycle: "true"
----
-kind: ConfigMap
-apiVersion: v1
-metadata:
-  name: connector-catalog-camel-misc
-  annotations:
-    qontract.recycle: "true"
----
-kind: ConfigMap
-apiVersion: v1
-metadata:
-  name: connector-catalog-debezium-mongodb
-  annotations:
-    qontract.recycle: "true"
----
-kind: ConfigMap
-apiVersion: v1
-metadata:
-  name: connector-catalog-debezium-mysql
-  annotations:
-    qontract.recycle: "true"
----
-kind: ConfigMap
-apiVersion: v1
-metadata:
-  name: connector-catalog-debezium-postgres
-  annotations:
-    qontract.recycle: "true"
